+++ driver/web/adapter.go
@@ -394,6 +394,21 @@ func (we Adapter) userAccountsAuthWrapFunc(handler userAccountsAuthFunc) http.Ha
 	}
 }
 
+type externalKeyAuthFunc = func(string, http.ResponseWriter, *http.Request)
+
+func (we Adapter) externalKeyAuthFunc(handler apiKeysAuthFunc) http.HandlerFunc {
+	return func(w http.ResponseWriter, req *http.Request) {
+		utils.LogRequest(req)
+
+		clientID, authenticated := we.auth.externalAuthCheck(w, req)
+		if !authenticated {
+			return
+		}
+
+		handler(clientID, w, req)
+	}
+}
+
 type loginAppUserRequest struct {
 	UUID                 string  `json:"uuid" validate:"required"`
 	PublicKey            string  `json:"public_key" validate:"required"`